建立 iOS 7 ARMv7 NativeAOT 构建原型

+++ src/coreclr/nativeaot/Runtime/unix/NativeContext.h
@@ -125,22 +125,22 @@ struct NATIVE_CONTEXT
         lambda((size_t*)&R15());
     }
 #elif defined(TARGET_ARM)
-    uint64_t& Pc();
-    uint64_t& Sp();
-    uint64_t& Lr();
-    uint64_t& R0();
-    uint64_t& R1();
-    uint64_t& R2();
-    uint64_t& R3();
-    uint64_t& R4();
-    uint64_t& R5();
-    uint64_t& R6();
-    uint64_t& R7();
-    uint64_t& R8();
-    uint64_t& R9();
-    uint64_t& R10();
-    uint64_t& R11();
-    uint64_t& R12();
+    uint32_t& Pc();
+    uint32_t& Sp();
+    uint32_t& Lr();
+    uint32_t& R0();
+    uint32_t& R1();
+    uint32_t& R2();
+    uint32_t& R3();
+    uint32_t& R4();
+    uint32_t& R5();
+    uint32_t& R6();
+    uint32_t& R7();
+    uint32_t& R8();
+    uint32_t& R9();
+    uint32_t& R10();
+    uint32_t& R11();
+    uint32_t& R12();
 
     uintptr_t GetIp() { return (uintptr_t)Pc(); }
     uintptr_t GetSp() { return (uintptr_t)Sp(); }
